requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Password hashing
bcrypt>=4.0.1
passlib>=1.7.4  # Added passlib for password hashing

# Database drivers (uncomment as needed)
# psycopg2-binary>=2.9.9  # PostgreSQL
# pymysql>=1.1.0  # MySQL
# aiosqlite>=0.19.0  # SQLite with async support
sqlalchemy-utils>=0.41.1

# Development tools
pytest>=7.4.0
httpx>=0.24.1  # For testing